+++ data-security-parent/data-security-core/src/main/java/io/telicent/smart/cache/security/data/labels/DefaultLabelApplicator.java
@@ -27,8 +27,12 @@ import org.apache.jena.sparql.core.Quad;
 @AllArgsConstructor
 public final class DefaultLabelApplicator implements SecurityLabelsApplicator {
 
-    @NonNull
     private final SecurityLabels<?> defaultLabel;
+
+    @Override
+    public SecurityLabels<?> defaultLabel() {
+        return this.defaultLabel;
+    }
 
     @Override
     public SecurityLabels<?> labelForTriple(Triple triple) {