+++ event-sources/event-source-kafka/src/main/java/io/telicent/smart/cache/sources/kafka/KafkaEvent.java
@@ -72,7 +72,14 @@ public class KafkaEvent<TKey, TValue> implements Event<TKey, TValue> {
     @Override
     public String lastHeader(String key) {
         List<String> values = this.headers(key).collect(Collectors.toList());
-        return CollectionUtils.isEmpty(values) ? null : values.get(values.size() - 1);
+        return CollectionUtils.isEmpty(values) ? null : values.getLast();
+    }
+
+    @Override
+    public EventHeader lastRawHeader(String key) {
+        return StreamSupport.stream(record.headers().headers(key).spliterator(), false)
+                            .map(h -> new RawHeader(h.key(), h.value()))
+                            .reduce(null, (r, e) -> e);
     }
 
     @Override